requirements.txt
Flask==3.1.0
openpyxl==3.1.5
matplotlib==3.10.0
mystic==0.4.3
numpy==2.2.1
pandas==2.2.3
scipy==1.15.1
pyinstaller==6.11.1
pywebview==5.3.2
waitress==3.0.2
pillow~=11.1.0
zlib~=1.2.13
pip~=24.2
wheel~=0.44.0
openssl~=3.0.15
pox~=0.3.5
cffi~=1.17.1
setuptools~=75.1.0
dill~=0.3.9
pytz~=2024.2
markupsafe~=3.0.2
click~=8.1.8
jinja2~=3.1.5
blinker~=1.9.0
itsdangerous~=2.2.0
mpmath~=1.3.0
sympy~=1.13.3
python~=3.12.8
pythonnet~=3.0.5
klepto~=0.2.6
python-dateutil~=2.9.0.post0
bottle~=0.13.2
proxy-tools~=0.1.0
altgraph~=0.17.4
six~=1.17.0
et-xmlfile~=2.0.0
contourpy~=1.3.1
fonttools~=4.55.3
packaging~=24.2
pefile~=2023.2.7
pycparser~=2.22
pyparsing~=3.2.1
clr-loader~=0.2.7.post0
cycler~=0.12.1
kiwisolver~=1.4.8
pywin32-ctypes~=0.2.3
pyinstaller-hooks-contrib~=2025.0
requests~=2.32.3